// src/app/api/model/firebase-auth-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FirebaseError } from './firebase-error';


export interface FirebaseAuthResponse { 
    correlationId?: string;
    displayName?: string;
    email?: string;
    emailVerified?: boolean;
    error?: FirebaseError;
    expiresIn?: string;
    idToken?: string;
    localId?: string;
    message?: string;
    processingTime?: number;
    refreshToken?: string;
    registered?: boolean;
    requires2FA?: boolean;
    requires2FASetup?: boolean;
    /**
     * User role from Firebase custom claims
     */
    role?: FirebaseAuthResponseRoleEnum;
    success?: boolean;
}
export enum FirebaseAuthResponseRoleEnum {
    ADMIN = 'ADMIN',
    PENDING_ADMIN = 'PENDING_ADMIN',
    INFLUENCER = 'INFLUENCER',
    COMPANY = 'COMPANY'
};