// src/app/api/model/token-exchange-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenExchangeResponse { 
    /**
     * Session cookie type
     */
    cookieType?: TokenExchangeResponseCookieTypeEnum;
    email?: string;
    error?: string;
    firebaseUid?: string;
    requires2FA?: boolean;
    requires2FASetup?: boolean;
    /**
     * User role (Permission/UserType name)
     */
    role?: TokenExchangeResponseRoleEnum;
    sessionDuration?: string;
    success?: boolean;
    twoFactorEnabled?: boolean;
    twoFactorVerified?: boolean;
    userId?: number;
}
export enum TokenExchangeResponseCookieTypeEnum {
    PARTIAL = 'PARTIAL',
    FULL = 'FULL'
};
export enum TokenExchangeResponseRoleEnum {
    ADMIN = 'ADMIN',
    PENDING_ADMIN = 'PENDING_ADMIN',
    INFLUENCER = 'INFLUENCER',
    COMPANY = 'COMPANY'
};